test: verify simulate tx handles error
